requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
opencv-python
icrawler
pillow